[skip ci] refinement (#37)
