+++ trisicell/tl/solver/huntress/__init__.py
@@ -1,53 +1,44 @@
 import pandas as pd
 
 import trisicell as tsc
-from trisicell.tl.solver.huntress._huntress_21_03_19 import Reconstruct
+from trisicell.tl.solver.huntress._huntress import Reconstruct
 
 
-def huntress(df_input_filepath, alpha, beta, kind="dna", n_threads=1):
+def huntress(df_input, alpha, beta, kind="both", n_threads=1):
     tsc.logg.info(
         f"running HUNTRESS with alpha={alpha}, beta={beta}, kind={kind},"
         f" n_threads={n_threads}"
     )
-    tmpdir = tsc.ul.tmpdirsys(suffix=".huntress")
 
     running_time = 0
-    if kind == "dna":
+    if kind == "both":
         fn_conorm = 0.1
         fp_conorm = fn_conorm * alpha / beta
-        # fnfp_conorm = fn_conorm / fp_conorm
-
-        running_time = Reconstruct(
-            df_input_filepath,
-            f"{tmpdir.name}/huntress",
+        matrix_out, running_time = Reconstruct(
+            df_input,
             n_proc=n_threads,
             fnfp=51,
             post_fn=fn_conorm,
             post_fp=fp_conorm,
         )
-        df_output = pd.read_table(f"{tmpdir.name}/huntress_optH_.CFMatrix", index_col=0)
     elif kind == "fn":
-        running_time = Reconstruct(
-            df_input_filepath,
-            f"{tmpdir.name}/huntress.CFMatrix",
+        matrix_out, running_time = Reconstruct(
+            df_input,
             Algchoice="FN",
             n_proc=n_threads,
         )
-        df_output = pd.read_table(f"{tmpdir.name}/huntress.CFMatrix", index_col=0)
     elif kind == "rna":
-        running_time = Reconstruct(
-            df_input_filepath,
-            f"{tmpdir.name}/huntress",
+        matrix_out, running_time = Reconstruct(
+            df_input,
             postprocessing=1,
             n_proc=n_threads,
         )
-        df_output = pd.read_table(
-            f"{tmpdir.name}/huntress_optH_.processed.CFMatrix", index_col=0
-        )
 
-    df_input = pd.read_table(df_input_filepath, index_col=0)
+    df_output = pd.DataFrame(matrix_out)
+    df_output.columns = df_input.columns
+    df_output.index = df_input.index
+    df_output.index.name = "cellIDxmutID"
+
     tsc.ul.stat(df_input, df_output, alpha, beta, running_time)
-
-    tmpdir.cleanup()
 
     return df_output
